facilitate async middleware and behaviors
the current endpoint management expectation is that each endpoint
does not have middleware. This update will allow each endpoint to
manage its own middleware, if needed.

This also means there's an opportunity to allow for configuration
of app wide middleware that aren't standard for the app.

as a last note, there's likely some variable naming that needs to
change here.