provide an interface for passing in a config and generating an app
makeApp is probably a method name that should migrate. The general
idea is that bin/www, on invocation, will receive a config object
or read it from variable, and pass it to app in order to create
the application.

This should allow the app to manage the utilities and invert control
from the boundary methods. In order for this to work as intended, the
makeRouter method will need to be adjusted to allow for passing in
dependencies into the behavior.

Some work will need to be done in order to avoid breaking middlewares.