process resources and create a route handler for each one
in order to provide programatic access to resources, this needs to
be managed differently, but the main goal at present is to port
functionality migrate a few services to this boilerplate.

This build will fail unless resource is defined by an import. In
addition, the makeRouter api will need to be refactored.