enable leveraging the boundary and entityMap
both the boundary and entity are true utils--the boundary could be
applied as middleware instead of as component logic, and some apps
likely do not need an entityMap. In the future, messaging components
will be provided as well. These methods should be provided to the
boundary behavior to be leveraged or not.